# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Asteroid Watch API - NASA NEO monitoring
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DiameterRange {
  estimated_diameter_min: number;
  estimated_diameter_max: number;
}

export interface EstimatedDiameter {
  kilometers: DiameterRange;
  meters: DiameterRange;
  miles: DiameterRange;
  feet: DiameterRange;
}

export interface RelativeVelocity {
  kilometers_per_second: string;
  kilometers_per_hour: string;
  miles_per_hour: string;
}

export interface MissDistance {
  astronomical: string;
  lunar: string;
  kilometers: string;
  miles: string;
}

export interface CloseApproachData {
  close_approach_date: string;
  close_approach_date_full: string;
  epoch_date_close_approach: number;
  relative_velocity: RelativeVelocity;
  miss_distance: MissDistance;
  orbiting_body: string;
}

export interface Neo {
  id: string;
  neo_reference_id: string;
  name: string;
  name_limited?: string;
  designation?: string;
  nasa_jpl_url: string;
  absolute_magnitude_h: number;
  estimated_diameter: EstimatedDiameter;
  is_potentially_hazardous_asteroid: boolean;
  close_approach_data: CloseApproachData[];
  is_sentry_object: boolean;
}

export interface NeoFeedDay {
  date: string;
  asteroids: Neo[];
}

export interface NeoFeed {
  element_count: number;
  start_date: string;
  end_date: string;
  near_earth_objects: NeoFeedDay[];
}

export interface TodayStats {
  asteroid_count: number;
  hazardous_count: number;
  /** @nullable */
  closest_approach_km: number | null;
  /** @nullable */
  fastest_velocity_kms: number | null;
  /** @nullable */
  largest_diameter_km: number | null;
  featured_asteroid: Neo | null;
}

export interface NasaStats {
  neo_count: number;
  close_approach_count: number;
  last_updated: string;
}

export interface NeoHistoryDay {
  date: string;
  total_count: number;
  hazardous_count: number;
  avg_velocity_kms: number;
  min_miss_distance_km: number;
  max_diameter_km: number;
}

export type GetNeoFeedParams = {
/**
 * Start date YYYY-MM-DD
 */
start_date: string;
/**
 * End date YYYY-MM-DD
 */
end_date: string;
};

export type SearchNeosParams = {
/**
 * Asteroid name or NASA designation
 */
query: string;
/**
 * Filter to potentially hazardous only
 */
hazardous_only?: boolean;
};