# lib/api-zod/src/generated/types/closeApproachData.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Asteroid Watch API - NASA NEO monitoring
 * OpenAPI spec version: 0.1.0
 */
import type { MissDistance } from './missDistance';
import type { RelativeVelocity } from './relativeVelocity';

export interface CloseApproachData {
  close_approach_date: string;
  close_approach_date_full: string;
  epoch_date_close_approach: number;
  relative_velocity: RelativeVelocity;
  miss_distance: MissDistance;
  orbiting_body: string;
}