# lib/api-zod/src/generated/types/diameterRange.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Asteroid Watch API - NASA NEO monitoring
 * OpenAPI spec version: 0.1.0
 */

export interface DiameterRange {
  estimated_diameter_min: number;
  estimated_diameter_max: number;
}