# lib/api-zod/src/generated/types/estimatedDiameter.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Asteroid Watch API - NASA NEO monitoring
 * OpenAPI spec version: 0.1.0
 */
import type { DiameterRange } from './diameterRange';

export interface EstimatedDiameter {
  kilometers: DiameterRange;
  meters: DiameterRange;
  miles: DiameterRange;
  feet: DiameterRange;
}