# lib/api-zod/src/generated/types/nasaStats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Asteroid Watch API - NASA NEO monitoring
 * OpenAPI spec version: 0.1.0
 */

export interface NasaStats {
  neo_count: number;
  close_approach_count: number;
  last_updated: string;
}