# lib/api-zod/src/generated/types/neo.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Asteroid Watch API - NASA NEO monitoring
 * OpenAPI spec version: 0.1.0
 */
import type { CloseApproachData } from './closeApproachData';
import type { EstimatedDiameter } from './estimatedDiameter';

export interface Neo {
  id: string;
  neo_reference_id: string;
  name: string;
  name_limited?: string;
  designation?: string;
  nasa_jpl_url: string;
  absolute_magnitude_h: number;
  estimated_diameter: EstimatedDiameter;
  is_potentially_hazardous_asteroid: boolean;
  close_approach_data: CloseApproachData[];
  is_sentry_object: boolean;
}