# lib/api-zod/src/generated/types/neoFeed.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Asteroid Watch API - NASA NEO monitoring
 * OpenAPI spec version: 0.1.0
 */
import type { NeoFeedDay } from './neoFeedDay';

export interface NeoFeed {
  element_count: number;
  start_date: string;
  end_date: string;
  near_earth_objects: NeoFeedDay[];
}