# lib/api-zod/src/generated/types/neoFeedDay.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Asteroid Watch API - NASA NEO monitoring
 * OpenAPI spec version: 0.1.0
 */
import type { Neo } from './neo';

export interface NeoFeedDay {
  date: string;
  asteroids: Neo[];
}