# lib/api-zod/src/generated/types/neoHistoryDay.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Asteroid Watch API - NASA NEO monitoring
 * OpenAPI spec version: 0.1.0
 */

export interface NeoHistoryDay {
  date: string;
  total_count: number;
  hazardous_count: number;
  avg_velocity_kms: number;
  min_miss_distance_km: number;
  max_diameter_km: number;
}