# lib/api-zod/src/generated/types/relativeVelocity.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Asteroid Watch API - NASA NEO monitoring
 * OpenAPI spec version: 0.1.0
 */

export interface RelativeVelocity {
  kilometers_per_second: string;
  kilometers_per_hour: string;
  miles_per_hour: string;
}