# lib/api-zod/src/generated/types/todayStats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Asteroid Watch API - NASA NEO monitoring
 * OpenAPI spec version: 0.1.0
 */
import type { Neo } from './neo';

export interface TodayStats {
  asteroid_count: number;
  hazardous_count: number;
  /** @nullable */
  closest_approach_km: number | null;
  /** @nullable */
  fastest_velocity_kms: number | null;
  /** @nullable */
  largest_diameter_km: number | null;
  featured_asteroid: Neo | null;
}